Move the archive tasks to a dedicated package
