+++ plugin/src/main/groovy/org/openbakery/XcodePlugin.groovy
@@ -27,6 +27,9 @@ import org.openbakery.appledoc.AppledocTask
 import org.openbakery.appstore.AppstorePluginExtension
 import org.openbakery.appstore.AppstoreUploadTask
 import org.openbakery.appstore.AppstoreValidateTask
+import org.openbakery.archiving.XcodeBuildArchiveTask
+import org.openbakery.archiving.XcodeBuildArchiveTaskIosAndTvOS
+import org.openbakery.archiving.XcodeBuildLegacyArchiveTask
 import org.openbakery.carthage.CarthageCleanTask
 import org.openbakery.carthage.CarthageUpdateTask
 import org.openbakery.cocoapods.CocoapodsBootstrapTask
