+++ plugin/src/test/groovy/org/openbakery/XcodeBuildArchiveTaskOSXSpecification.groovy
@@ -4,6 +4,7 @@ import org.apache.commons.configuration.plist.XMLPropertyListConfiguration
 import org.apache.commons.io.FileUtils
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
+import org.openbakery.archiving.XcodeBuildLegacyArchiveTask
 import org.openbakery.testdouble.PlistHelperStub
 import org.openbakery.xcode.Type
 import spock.lang.Specification
